Move updatable scheduled task types to /models
