+++ src/client.spec.ts
@@ -27,8 +27,6 @@ import {
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
-	UpdatableScheduledQuery,
-	UpdatableScheduledScript,
 	UpdatableTemplate,
 	UpdatableUser,
 	ValidatedQuery,
@@ -88,6 +86,8 @@ import {
 	UpdatablePlaybook,
 	UpdatableResource,
 	UpdatableSavedQuery,
+	UpdatableScheduledQuery,
+	UpdatableScheduledScript,
 	UploadableAutoExtractor,
 	User,
 	UserPreferences,
