+++ src/functions/scheduled-tasks/update-one-scheduled-query.ts
@@ -6,9 +6,9 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { ScheduledQuery } from '../../models';
+import { ScheduledQuery, UpdatableScheduledQuery } from '../../models';
 import { APIFunctionMakerOptions } from '../utils';
-import { makeUpdateOneScheduledTask, UpdatableScheduledQuery } from './update-one-scheduled-task';
+import { makeUpdateOneScheduledTask } from './update-one-scheduled-task';
 
 export const makeUpdateOneScheduledQuery = (makerOptions: APIFunctionMakerOptions) => {
 	const updateOneScheduleTask = makeUpdateOneScheduledTask(makerOptions);
