+++ src/models/scheduled-task/index.ts
@@ -15,12 +15,15 @@ export * from './is-scheduled-script';
 export * from './is-scheduled-task-base';
 export * from './raw-creatable-scheduled-task';
 export * from './raw-scheduled-task';
+export * from './scheduled-query';
 export * from './scheduled-script';
 export * from './scheduled-task';
 export * from './scheduled-task-base';
-export * from './squeduled-query';
 export * from './tagged-creatable-scheduled-query';
 export * from './tagged-creatable-scheduled-script';
 export * from './to-raw-creatable-scheduled-task';
+export * from './to-raw-updatable-scheduled-task';
 export * from './to-scheduled-task';
 export * from './to-scheduled-task-base';
+export * from './updatable-scheduled-task';
+export * from './updatable-scheduled-task-base';
