+++ src/models/scheduled-task/scheduled-query.ts
